api-gateway/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.28,<1.0
python-jose[cryptography]>=3.3,<4.0
pytest>=8.3,<9.0
black>=25.1,<26.0
ruff>=0.11,<1.0
